Replaces the storage type Float with Double (SE-0307)

+++ Sources/Swift2D/Point+Computed.swift
@@ -1,9 +1,9 @@
 // MARK: - Static References
 public extension Point {
     static let zero: Point = Point(x: 0, y: 0)
-    static let nan: Point = Point(x: Float.nan, y: Float.nan)
-    internal static let infinite: Point = Point(x: -Float.greatestFiniteMagnitude / 2, y: -Float.greatestFiniteMagnitude / 2)
-    internal static let null: Point = Point(x: Float.infinity, y: Float.infinity)
+    static let nan: Point = Point(x: Double.nan, y: Double.nan)
+    internal static let infinite: Point = Point(x: -Double.greatestFiniteMagnitude / 2, y: -Double.greatestFiniteMagnitude / 2)
+    internal static let null: Point = Point(x: Double.infinity, y: Double.infinity)
     
     var isZero: Bool {
         return self == .zero
